requirements.txt
gaiatest-v2.0
pytest-mozwebqa
unittestzero
